Remove releaseList to its own file
