+++ rpc/releaselist.go
@@ -1,0 +1,14 @@
+package rpc
+
+import "capnproto.org/go/capnp/v3"
+
+type releaseList []capnp.ReleaseFunc
+
+func (rl releaseList) release() {
+	for _, r := range rl {
+		r()
+	}
+	for i := range rl {
+		rl[i] = func() {}
+	}
+}
